Do line ending normalization by default again
