+++ src/test/java/com/diffplug/gradle/spotless/ResourceTest.java
@@ -33,7 +33,7 @@ import org.junit.rules.TemporaryFolder;
 
 import com.diffplug.common.base.Throwing;
 
-public class ResourceTest {
+public abstract class ResourceTest {
 	@Rule
 	public TemporaryFolder folder = new TemporaryFolder();
 
@@ -82,34 +82,4 @@ public class ResourceTest {
 		Assert.assertEquals(expected, formatted);
 	}
 
-	/** Creates a FormatTask based on the given consumer. */
-	public static FormatTask createTask(Consumer<FormatExtension> test) throws Exception {
-		Project project = ProjectBuilder.builder().build();
-		SpotlessPlugin plugin = project.getPlugins().apply(SpotlessPlugin.class);
-
-		AtomicReference<FormatExtension> ref = new AtomicReference<>();
-		plugin.getExtension().format("underTest", ext -> {
-			ref.set(ext);
-			test.accept(ext);
-		});
-
-		boolean check = false;
-		return plugin.createTask("underTest", ref.get(), check);
-	}
-
-	/** Tests that the formatExtension causes the given change. */
-	protected void assertTask(Consumer<FormatExtension> test, String before, String afterExpected) throws Exception {
-		// create the task
-		FormatTask task = createTask(test);
-		// create the test file
-		File testFile = folder.newFile();
-		Files.write(testFile.toPath(), before.getBytes(StandardCharsets.UTF_8));
-		// set the task to use this test file
-		task.target = Collections.singleton(testFile);
-		// run the task
-		task.format();
-		// check what the task did
-		String afterActual = new String(Files.readAllBytes(testFile.toPath()), StandardCharsets.UTF_8);
-		Assert.assertEquals(afterExpected, afterActual);
-	}
 }
